requirements.txt
fastapi
tortoise-orm
python-multipart
aiosqlite
aiofiles
jinja2